requirements.txt
pillow
onnxruntime